Aqi_website_code/requirements.txt
fastapi
uvicorn
pandas
pymongo
scikit-learn
python-dotenv
pydantic
numpy
